# .ourent/checkpoints/2026-10-08T06-41-31-984Z-8f7ec244/files/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kamol Aliboyev Portfolio API
  version: 1.0.0
  description: |
    REST API for the portfolio, store, blog, comments, cart, orders, payments and admin panel.
    Every response uses an envelope: `{success, data, pagination?}` or `{success:false, message, error:{code, fields?}}`.
    Authentication: short-lived Bearer access token plus an HttpOnly refresh cookie (`kamol_rt`).
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearer: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    Error:
      type: object
      properties:
        success: {type: boolean, example: false}
        message: {type: string}
        error: {type: object, properties: {code: {type: string, example: PRODUCT_NOT_FOUND}, fields: {type: object, additionalProperties: {type: string}}}}
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: "Create an account and start a session"
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
            example: {"name": "Aziza Karimova", "email": "aziza@example.com", "password": "S3cure-passw0rd", "confirm_password": "S3cure-passw0rd"}
      responses:
        "201": {description: Created}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/login:
    post:
      tags: [Auth]
      summary: "Log in (sets the HttpOnly refresh cookie, returns a short-lived access token)"
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
            example: {"email": "aziza@example.com", "password": "S3cure-passw0rd"}
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/logout:
    post:
      tags: [Auth]
      summary: "Revoke the refresh token family and clear the cookie"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/refresh:
    post:
      tags: [Auth]
      summary: "Rotate the refresh cookie and get a new access token"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/me:
    get:
      tags: [Auth]
      summary: "Current user"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: "Request a password reset email (always answers 200)"
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
            example: {"email": "aziza@example.com"}
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: "Set a new password with a reset token"
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
            example: {"token": "<token from email>", "password": "New-passw0rd-123"}
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /me:
    get:
      tags: [Users]
      summary: "Profile"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    patch:
      tags: [Users]
      summary: "Update name, avatar or password"
      security: [{bearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object}
            example: {"name": "Aziza K."}
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /me/orders:
    get:
      tags: [Users]
      summary: "My orders (paginated)"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /me/purchases:
    get:
      tags: [Users]
      summary: "Products I have access to"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /me/favorites:
    get:
      tags: [Users]
      summary: "My favorite products"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /products:
    get:
      tags: [Products]
      summary: "List published products (search, category, featured, sort, page, limit)"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /products/{ref}:
    parameters:
      - {name: ref, in: path, required: true, schema: {type: string}}
    get:
      tags: [Products]
      summary: "Product by id or slug"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /products/{id}/access:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    get:
      tags: [Products]
      summary: "Check purchase access and get a 5-minute download URL"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /products/{id}/download:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    get:
      tags: [Products]
      summary: "302 redirect to the signed download URL (purchasers only)"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /products/{id}/favorite:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    post:
      tags: [Favorites]
      summary: "Add to favorites"
      security: [{bearer: []}]
      responses:
        "201": {description: Created}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    delete:
      tags: [Favorites]
      summary: "Remove from favorites"
      security: [{bearer: []}]
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "401": {description: Authentication required, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /projects:
    get:
      tags: [Projects]
      summary: "List published projects"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /projects/{ref}:
    parameters:
      - {name: ref, in: path, required: true, schema: {type: string}}
    get:
      tags: [Projects]
      summary: "Project by id or slug (includes like_count / liked)"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /posts:
    get:
      tags: [Blog]
      summary: "List published posts (search, category, sort, page, limit)"
      responses:
        "200": {description: OK}
        "400": {description: Malformed request, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: Not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: Conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: Validation error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "429": {description: Rate limited, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...